User creates an innovation
Any innovations will be tied to the logged in user account.  If the user
isn't logged in they'll be redirected to authenticate before allowing
the entry of an innovation.  When looking at all innovations if the user
has logged in they will just see the innovations they've provided (since
we have search functionality in place).

Validation:

- Log out and visit http://localhost:3000/
- Click "Innovations" (Header)
- Verify that all innovations are listed

- Click "New Innovation" (Header)
- Authenticate with GitHub
- Click "New Innovation" (Header)
- Enter valid data to create the innovation
- Click "Innovations" (Header)
- Verify that only the new innovation is present
- Log out

- Click "New Innovation" (Header)
- Authenticate with Google
- Repeat the steps noted above